requirements.txt
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4

djangorestframework~=3.14.0
python-dotenv~=1.0.0
django-filter~=23.5
Faker~=21.0.0